requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
azure-ai-vision-imageanalysis==1.0.0b2
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.30.2
black==24.4.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
Django==4.2.13
django-environ==0.11.2
djangorestframework==3.15.1
dnspython==2.6.1
eval-type-backport==0.1.3
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.3
idna==3.7
ipykernel==6.29.4
ipython==8.25.0
isodate==0.6.1
jedi==0.19.1
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-path==0.3.2
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy-object-proxy==1.10.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
more-itertools==10.2.0
motor==3.4.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.32.0
openapi-core==0.19.1
openapi-schema-validator==0.6.2
openapi-spec-validator==0.7.1
opencv-python==4.10.0.82
packaging==24.0
parse==1.20.1
parso==0.8.4
pathable==0.4.3
pathspec==0.12.1
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.2
prance==23.6.21.0
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pybars4==0.9.13
pycountry==24.6.1
pycparser==2.22
pydantic==2.7.3
pydantic_core==2.18.4
PyGithub==2.3.0
Pygments==2.18.0
PyJWT==2.8.0
PyMeta3==0.5.1
pymongo==4.7.3
PyNaCl==1.5.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.31.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rpds-py==0.18.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
safetensors==0.4.3
scipy==1.13.1
semantic-kernel==0.9.6b1
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
sympy==1.12.1
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
torch==2.2.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.41.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
